Override the current gmres method and move solve! to A_ldiv_B!
